pyproject.toml
[project]
name = "streamlens"
version = "0.1.0"
description = "Analyze video content from streaming platforms like Bilibili"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "yt-dlp",
    "openai-whisper",
]

[project.scripts]
streamlens = "streamlens.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
